some additional code reorg for faster browser test automation authoring
