+++ documentation/library_list.md
@@ -87,6 +87,11 @@ This is a dynamically compiled list of supporting code files that comprise this 
 * Directory *[../lib/terminal/test/application/browserUtilities](../lib/terminal/test/application/browserUtilities)*
    - **[../lib/terminal/test/application/browserUtilities/file_path_decode.ts](../lib/terminal/test/application/browserUtilities/file_path_decode.ts)** - Transforms a custom encoded file path into a local operation system specific file path.
    - **[../lib/terminal/test/application/browserUtilities/file_path_encode.ts](../lib/terminal/test/application/browserUtilities/file_path_encode.ts)** - Creates an encoding around file system addresses so that the test code can ensure the paths are properly formed per operating system.
+   - **[../lib/terminal/test/application/browserUtilities/inviteAccept.ts](../lib/terminal/test/application/browserUtilities/inviteAccept.ts)**         - A test generator for accepting an invitation.
+   - **[../lib/terminal/test/application/browserUtilities/inviteConfirm.ts](../lib/terminal/test/application/browserUtilities/inviteConfirm.ts)**       - A test generator for accepting an invitation.
+   - **[../lib/terminal/test/application/browserUtilities/inviteModal.ts](../lib/terminal/test/application/browserUtilities/inviteModal.ts)**           - A test generator for spawning the invitation modal.
+   - **[../lib/terminal/test/application/browserUtilities/inviteSend.ts](../lib/terminal/test/application/browserUtilities/inviteSend.ts)**             - A test generator for sending an invitation.
+   - **[../lib/terminal/test/application/browserUtilities/login.ts](../lib/terminal/test/application/browserUtilities/login.ts)**                       - A test generator to login a fresh device.
    - **[../lib/terminal/test/application/browserUtilities/machines.ts](../lib/terminal/test/application/browserUtilities/machines.ts)**                 - An object expressing a map of the various remote machines.
    - **[../lib/terminal/test/application/browserUtilities/mainMenu.ts](../lib/terminal/test/application/browserUtilities/mainMenu.ts)**                 - A convenience function that opens the main menu while in browser tests.
    - **[../lib/terminal/test/application/browserUtilities/modalAddress.ts](../lib/terminal/test/application/browserUtilities/modalAddress.ts)**         - A convenience function that tests a file navigation modal to go to the project's location for browser tests.
