+++ lib/terminal/test/application/browserUtilities/readme.md
@@ -6,6 +6,11 @@ The convenience functions for generating various browser tests.
 
 * **[file_path_decode.ts](file_path_decode.ts)** - Transforms a custom encoded file path into a local operation system specific file path.
 * **[file_path_encode.ts](file_path_encode.ts)** - Creates an encoding around file system addresses so that the test code can ensure the paths are properly formed per operating system.
+* **[inviteAccept.ts](inviteAccept.ts)**         - A test generator for accepting an invitation.
+* **[inviteConfirm.ts](inviteConfirm.ts)**       - A test generator for accepting an invitation.
+* **[inviteModal.ts](inviteModal.ts)**           - A test generator for spawning the invitation modal.
+* **[inviteSend.ts](inviteSend.ts)**             - A test generator for sending an invitation.
+* **[login.ts](login.ts)**                       - A test generator to login a fresh device.
 * **[machines.ts](machines.ts)**                 - An object expressing a map of the various remote machines.
 * **[mainMenu.ts](mainMenu.ts)**                 - A convenience function that opens the main menu while in browser tests.
 * **[modalAddress.ts](modalAddress.ts)**         - A convenience function that tests a file navigation modal to go to the project's location for browser tests.
